requirements.txt
pandas
numpy
requests
python-telegram-bot>=20.0
pytest
pandas-ta
